initial spike that sort of works, albeit badly
No commonjs support yet, and it's way too slow. And I'm not entirely
clear if it's actually type checking or not? It seems to allow anything
through even when there are clearly type errors.

Profiler indicates that a huge chunk of the time is being spent in
`makeSyncRequest` in the loader subsystem, so I'm next going to try
doing the transpilation directly on the loader thread instead of on the
main thread.

That might mean that the CommonJS require.extensions hooks will end up
being duplicative, but since we only transpile code in the current
project, that would only be an issue in cases where the project has a
mix of CJS and ESM, and it's loaded using `import`, if I'm correctly
understanding how all this stuff works. (Big "if".)

Since require.extensions can't be async, that also means the whole
client/service bits are going away. Might find some other use for it,
maybe not though. It was mostly copied over from @tapjs/mock anyway, so
no great loss.